chore: Update build.sh to improve TRY and Install modes
This commit updates the build.sh file to improve the TRY and Install modes in the generated grub.cfg. It enforces the username and hostname as "ubuntu", enforces X11, and addresses issues with logging out and locking the screen. In TRY mode, it adds an "Install" icon for the Ubiquity installer, while in Install mode, it disables all gnome extensions and shows the Ubiquity installer by default. After installation, it enables the desktop, all gnome extensions, and removes the Ubiquity installer and "Install" icon. Additionally, it clarifies the behavior for Nvidia users and other users regarding X11 and Wayland usage.