chore: Upgrade system to 0.2.1-beta
This commit adds the necessary upgrade steps from version 0.2.0 to version 0.2.1-beta. It installs new packages, patches localization, and handles the installation of a new plugin. Additionally, it updates the system to the latest version and restarts the ibus engine if ibus-rime is installed. This upgrade ensures that the system is up to date and includes the latest features and improvements.